+++ Assignment_5/graphicsLab/project/data/TAL.vert
@@ -31,7 +31,7 @@ varying lowp vec4 diffuseVarying;
 varying lowp vec4 specularVarying;
 varying lowp vec4 texCoordVarying;
 
-varying mediump vec4 posVarying;       // pos in world space
+varying highp vec4 posVarying;       // pos in world space
 varying mediump vec3 normalVarying;    // normal in world space
 
 varying highp vec3 wsEyePosition;
